+++ open_manipulator_bringup/scripts/init_position_for_x.py
@@ -1,0 +1,52 @@
+#!/usr/bin/env python3
+#
+# Copyright 2024 ROBOTIS CO., LTD.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# Author: Sungho Woo
+
+import rclpy
+from rclpy.node import Node
+from trajectory_msgs.msg import JointTrajectory, JointTrajectoryPoint
+import time
+
+class MoveToHome(Node):
+    def __init__(self):
+        super().__init__('move_to_home')
+        self.publisher = self.create_publisher(JointTrajectory, '/arm_controller/joint_trajectory', 10)
+        self.timer = self.create_timer(1.0, self.move_to_home)
+
+    def move_to_home(self):
+        traj = JointTrajectory()
+        traj.joint_names = ['joint1', 'joint2', 'joint3', 'joint4']
+
+        point = JointTrajectoryPoint()
+        point.positions = [0.0, -1.0, 1.0, 0.0]
+        point.time_from_start.sec = 3
+
+        traj.points.append(point)
+        self.publisher.publish(traj)
+        self.get_logger().info('Moving to Home Position')
+
+        self.timer.cancel()
+        self.destroy_node()
+        rclpy.shutdown()
+
+def main(args=None):
+    rclpy.init(args=args)
+    node = MoveToHome()
+    rclpy.spin(node)
+
+if __name__ == '__main__':
+    main()